requirements.txt
amqp==2.6.1
appdirs==1.4.4
asyncio==3.4.3
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.0
cycler==0.10.0
distlib==0.3.1
Django==2.1
django-rest-knox==4.1.0
djangorestframework==3.11.1
dnspython==1.16.0
eventlet==0.26.1
filelock==3.0.12
Flask==1.1.2
Flask-SocketIO==4.3.1
gevent==20.6.2
greenlet==0.4.16
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
joblib==0.14.1
kafka==1.3.5
kafka-python==2.0.1
kiwisolver==1.1.0
kombu==4.6.11
MarkupSafe==1.1.1
matplotlib==3.0.3
monotonic==1.5
nltk==3.5
numpy==1.18.5
Pillow==7.2.0
pipenv==2020.6.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2020.1
redis==3.5.3
regex==2020.6.8
requests==2.24.0
scikit-learn==0.22.2.post1
scipy==1.2.1
six==1.15.0
soupsieve==2.0.1
stop-words==2018.7.23
tqdm==4.47.0
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.28
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wordcloud==1.7.0
zipp==1.2.0
zope.event==4.4
zope.interface==5.1.0
